Separating the 'beeb_' functions which are called from Swift into their own bridging body file
Added DiskImgs to the Resources (for the HardDrive)
