+++ .gitignore
@@ -74,3 +74,4 @@ fastlane/test_output
 # https://github.com/johnno1962/injectionforxcode
 
 iOSInjectionProject/
+BeebEm5.xcodeproj/project.xcworkspace/xcshareddata
