+++ Bugs.rtf
@@ -1,0 +1,58 @@
+{\rtf1\ansi\ansicpg1252\cocoartf2513
+\cocoatextscaling0\cocoaplatform0{\fonttbl\f0\fnil\fcharset0 LucidaGrande;}
+{\colortbl;\red255\green255\blue255;\red38\green38\blue38;\red231\green240\blue245;\red76\green123\blue175;
+\red16\green63\blue118;\red255\green255\blue255;}
+{\*\expandedcolortbl;;\cssrgb\c20000\c20000\c20000;\cssrgb\c92549\c95294\c96863;\cssrgb\c36471\c56078\c74118;
+\cssrgb\c6275\c32157\c53725;\cssrgb\c100000\c100000\c100000;}
+\paperw12240\paperh15840\margl1440\margr1440\vieww9000\viewh8400\viewkind0
+\deftab720
+\pard\pardeftab720\partightenfactor0
+
+\f0\fs26 \cf2 \cb3 \expnd0\expndtw0\kerning0
+\outl0\strokewidth0 \strokec2 Some testing of BeebEm5 for none game use on a MacBookPro running Catalina (10:15:7)\
+\
+Installation from here:\'a0{\field{\*\fldinst{HYPERLINK "https://github.com/CommanderCoder/BeebEm"}}{\fldrslt \cf4 \strokec4 \uc0\u8234 https://github.com/CommanderCoder/BeebEm}}\
+No issues with the.zip file, however, I've dropped the 3 files into a copy of my BeebEm4 (renamed to 5) folder that holds all my setup/configuration settings. First question: is this the correct way to set up?\
+\
+\cf0 \cb1 \kerning1\expnd0\expndtw0 \outl\strokewidth60 \strokec0 PROBABLY NOT:  The emulator expects to be in the 'applications' folder and when it executes it grabs files from the 'resources' folder within the application.\
+\
+This is likely undesired behaviour; if the user wants to have a different emulator for each folder of files, then this won't work.  However if the emulator is supposed to be self contained then this is expected.  I can revert to the one BEEBEM per user folder; or I can have it as an application which the use can select a FOLDER to start the application in.  Then changing this folder changes the setup easily.  Maybe even setup some shortcuts.\
+\
+\cf2 \cb3 \expnd0\expndtw0\kerning0
+\outl0\strokewidth0 \strokec2 \
+\
+There doesn't seem to be an up-to-date version number - or am I looking in the wrong place?\
+\pard\pardeftab720\partightenfactor0
+{\field{\*\fldinst{HYPERLINK "https://stardot.org.uk/forums/download/file.php?id=64569&mode=view"}}{\fldrslt \cf5 \cb6 \strokec5 \uc0\u8234 }}\cf5 \cb6 \strokec5 \
+\
+\pard\pardeftab720\partightenfactor0
+\cf0 \cb1 \kerning1\expnd0\expndtw0 \outl\strokewidth60 \strokec0 I'm not quite sure what the version number is ; it is somewhere between version 3 and version 4.\
+\pard\pardeftab720\partightenfactor0
+\cf5 \cb6 \expnd0\expndtw0\kerning0
+\outl0\strokewidth0 \strokec5 \
+\pard\pardeftab720\partightenfactor0
+\cf2 \cb1 \strokec2 \
+\pard\pardeftab720\partightenfactor0
+\cf2 \cb3 BeebEm5 setup as:\
+Master 128, Acorn 1772, 'ROM write' from 0 to 7 (this setting doesn't save!), Hard drive on - Configured to ADFS (File 13) - Speed set to 50 fps (I found this the most stable e.g. \
+dropping the least number of key presses.\
+\
+\cf0 \cb1 \kerning1\expnd0\expndtw0 \outl\strokewidth60 \strokec0 Hard Drive not tested\
+The version I used (version 3) didn't save ROM write settings!\cf2 \cb3 \expnd0\expndtw0\kerning0
+\outl0\strokewidth0 \strokec2 \
+\
+When you load a disc there's no indication that it's loaded! I mention this as once I load a .ADF disc then *. I get a "Broken directory" message - *MOUNT 0 doesn't produce an error message, however *. still produces "Broken directory". Changing to DFS (*DISC) then *. gives "Bad FS map".\
+\
+I have had a .adf disc running, however, it seems to be hit and miss! Another route. Configure to DFS - RESET (from the Menu) - Load a .adf disc to drive 0 - *ADFS - *. produces a "Disc fault 18 at 00/00" message. I haven't been able to mount a hard drive, as yet!\
+\
+I can configure to DFS - RESET - load a .ssd - *. and all is OK.\
+\
+*ROMS produces a list, as expected, however, changing the file "Roms.cfg" has no effect on the list.\
+\
+Missing keyboard entries. Typing at any speed I see dropped entries at random times.\
+\
+Speed setting don't seem to work and has a big influence on fps! Running a FOR-NEXT loop gives the same "TIME" at any speed! Setting the speed to 100 shows "Speed: (about) 1.xx fps: "1"\'a0\'a0\'a0. . .\'a0\'a0\'a0Setting the speed to 1.1 shows "Speed: (about) 1 fps: (around) 20.\
+\
+Until ADFS and hard drives are working I'm unable to do more testing. Anything you'd like me to test?\
+\
+Dave H.}
