+++ src/main.cc
@@ -41,12 +41,6 @@
 #include "speech.h"
 #include "main.h"
 
-struct CColour{
-    unsigned char r;
-    unsigned char g;
-    unsigned char b;
-    unsigned char a;
-};
 
 #if 0 //ACH  -push/pop symbolic hotkey
 
@@ -57,7 +51,6 @@ extern void PopSymbolicHotKeyMode(void * inToken)          __attribute__((weak_i
 FILE *tlog;
 int trace;
 int DumpAfterEach=0;
-int done = 0;
 #include "via.h"
 extern VIAState SysVIAState;
 BeebWin *mainWin;
@@ -133,11 +126,6 @@ static pascal OSErr AEodoc(const AppleEvent *theEvent, AppleEvent *theReply, lon
 
 char AutoRunPath[512]="";
 
-extern "C" void beeb_autorun(char* path)
-{
-    strcpy(AutoRunPath, path);
-}
-
 void AutoRunFromPath()
 {
     char* path = AutoRunPath;
@@ -200,104 +188,6 @@ CantGetParameter:
     return err;
 }
 #endif
-
-extern "C" void beeb_handlekeys(long eventkind, unsigned int keycode, char charCode)
-{
-     static int ctrl = 0x0000;
-     int LastShift, LastCtrl, LastCaps, LastCmd;
-     int NewShift, NewCtrl, NewCaps;
-     static int NewCmd = 0;
-    
-    switch (eventkind)
-    {
-        case kEventRawKeyDown:
-
-
-        //      fprintf(stderr, "Key pressed: code = %d, '%c'\n", keycode, charCode);
-                if ( (NewCmd) && (keycode == 6) )
-                {
-                    fprintf(stderr, "cmd-Z pressed, NewCmd = %d\n", NewCmd);
-                    if (mainWin->m_isFullScreen)
-                    {
-
- //                       EndFullScreen(mainWin->m_RestoreState, nil);
-                        mainWin->m_isFullScreen = 0;
-                        mainWin->SetMenuCommandIDCheck('vfsc', false);
-                    }
-                } else if ( (NewCmd) && (NewCtrl) && (keycode == 1) )
-                {
-                    fprintf(stderr, "ctrl-cmd-S pressed, NewCmd = %d\n", NewCmd);
-//                    mainWin->m_PrintScreen = true;
-                } else if ( (NewCmd) && (keycode == 8) )
-                {
-                    fprintf(stderr, "cmd-C pressed, NewCmd = %d\n", NewCmd);
-//                    mainWin->doCopy();
-                } else if ( (NewCmd) && (keycode == 9) )
-                {
-                    fprintf(stderr, "cmd-V pressed, NewCmd = %d\n", NewCmd);
-//                    mainWin->doPaste();
-                } else if ( (NewCmd) && (keycode == 17) )
-                {
-                    trace_186 = 1 - trace_186;
-                }
-              else
-                {
-                    mainWin->KeyDown(keycode);
-                }
-            break;
-          case kEventRawKeyUp:
-    //        fprintf(stderr, "Key released: code = %d, '%c'\n", keycode, charCode);
-            mainWin->KeyUp(keycode);
-            break;
-          case kEventRawKeyModifiersChanged:
-
-            /* Cocoa Event Handling used to be:
-             cmd 0x0000
-             shift 0x0200
-             alphalock 0x0400
-             option 0x0800
-             ctrl 0x1000
-             rightcmd 0x0001
-             rightshift 0x2000
-             rightoption 0x4000
-             rightctrl 0x8000
-             */
-            
-            /* new codes are (left/right or toggle):
-             cmd 0x100008
-             shift 0x20002
-             alphalock 0x10000 & 0x00000  - toggle on off - unused as keycodes from mac will be capslocked too
-             option 0x80020
-             ctrl 0x40001
-             rightshift 0x20004
-             rightoption 0x80040 - available for COPY
-             rightctrl 0x42001 - unused as only one ctrl key
-             rightcmd 0x100010 - available for COPY
-             fn  0x800000
-             
-             */
-                // ALPHALOCK key is unused
-            
-                LastShift = ctrl & 0x20006; // capture left and right shift
-                LastCtrl = ctrl & 0x40001; // capture left
-                LastCaps = ctrl & 0x80020; // capture left
-                LastCmd = ctrl & 0x100018; // capture left and right CMD
-
-                NewShift = keycode & 0x20006; // capture left and right shift
-                NewCtrl = keycode & 0x40001; // capture left
-                NewCaps = keycode & 0x80020; // capture left
-                NewCmd = keycode &  0x100018; // capture left and right CMD
-
-            
-//                fprintf(stderr, "Key modifier : code = %016x\n", keycode);
-                
-                if (LastShift != NewShift) if (LastShift) mainWin->KeyUp(200); else mainWin->KeyDown(200);
-                if (LastCtrl  != NewCtrl)  if (LastCtrl)  mainWin->KeyUp(201); else mainWin->KeyDown(201);
-                if (LastCaps  != NewCaps)  if (LastCaps)  mainWin->KeyUp(202); else mainWin->KeyDown(202);
-                ctrl = keycode;
-            break;
-    }
-}
 
 #if 0//ach - eventhandlers / keys&mouse, quit, main loop
 
@@ -662,174 +552,6 @@ static void RunApplicationEventLoopWithCooperativeThreadSupport(void)
 }
 
 #endif
-
-extern "C" void beeb_MainCpuLoop()
-{
-#if 0 //ACH- iswindowcollapsed
-    if ( (mainWin->m_FreezeWhenInactive) && (IsWindowCollapsed(mainWin->mWindow)) )
-    {
-        beeb_usleep(1000 * 500);        // sleep for 0.5 secs
-    }
-    else
-#endif
-    {
-        int c;
-        c = 20;
-
-    // Menu GUI more responsive if running less than real time
-        
-        if ( (mainWin->m_RealTimeTarget != 0) && (mainWin->m_RealTimeTarget < 1) )
-        {
-            c = c * mainWin->m_RealTimeTarget;
-        }
-        
-        for (int i = 0; i < c; ++i)
-            Exec6502Instruction();
-    }
-}
-int y = 0;
-const int width = 640;
-const int height = 512;
-
-
-const int bpp = 4; // 4 bytes or 32 bits
-unsigned char videobuffer[bpp * width * height]; // 200x200 by 4 bytes
-
-extern "C" void beeb_video(int height, int width, struct CColour buffer[])
-{
-    for (int i = 0; i < width*height; i++)
-    {
-        buffer[i].b = videobuffer[0+(i*bpp)];
-        buffer[i].g = videobuffer[1+(i*bpp)];
-        buffer[i].r = videobuffer[2+(i*bpp)];
-        buffer[i].a = videobuffer[3+(i*bpp)];
-    }
-    // move a green line everytime this is updated (50fps)
-    y += 1;
-    y %= height-2;
-}
-
-extern "C" int beeb_main(int argc,char *argv[])
-{
-void *token;
-int i;
-
-    // NEED TO TURN OFF SANDBOXING IN ENTITLEMENTS FILE TO GET LOCAL FOLDERS TO WORK
-    
-  fprintf(stderr, "Version: %s %s\n", Version, VersionDate);
-    char cwd[PATH_MAX];
-    if (getcwd(cwd, sizeof(cwd)) != NULL) {
-        printf("Current working dir: %s\n", cwd);
-    } else {
-        perror("getcwd() error");
-        return 1;
-    }
-
-
-  for (i = 0; i < argc; ++i)
-	fprintf(stderr, "Arg %d = %s\n", i, argv[i]);
-
-  mainWin=new BeebWin();
-#if 0 //ACH - main init hotkey
-
-  atexit(AtExitHandler);
-
-  if (PushSymbolicHotKeyMode != NULL)
-  {
-    token = PushSymbolicHotKeyMode(kHIHotKeyModeAllDisabled);
-  }
-
-
-  tlog = fopen("/users/jonwelch/trace.log", "wt");
-//  tlog = NULL;
-#endif
-    
-  done = 0;
-  
-  mainWin->Initialise(argv[0]);
-
-//  SoundReset();
-  if (SoundEnabled) SoundInit();
-  if (SpeechDefault) tms5220_start();
-  mainWin->ResetBeebSystem(MachineType,TubeEnabled,1); 
-  mainWin->SetRomMenu();
-  mainWin->SetSoundMenu();
-#if 0 //ACH - main init events
-
-  EventTypeSpec    eventTypes[7];
-  EventHandlerUPP  handlerUPP;
-
-  eventTypes[0].eventClass = kEventClassKeyboard;
-  eventTypes[0].eventKind  = 1; // kEventRawKeyDown
-  eventTypes[1].eventClass = kEventClassKeyboard;
-  eventTypes[1].eventKind  = 3; // kEventRawKeyUp
-  eventTypes[2].eventClass = kEventClassKeyboard;
-  eventTypes[2].eventKind  = 4; // kEventRawKeyModifiersChanged
-
-  eventTypes[3].eventClass = kEventClassMouse;
-  eventTypes[3].eventKind  = kEventMouseDown;
-  eventTypes[4].eventClass = kEventClassMouse;
-  eventTypes[4].eventKind  = kEventMouseUp;
-  eventTypes[5].eventClass = kEventClassMouse;
-  eventTypes[5].eventKind  = kEventMouseMoved;
-  eventTypes[6].eventClass = kEventClassMouse;
-  eventTypes[6].eventKind  = kEventMouseDragged;
-
-  handlerUPP = NewEventHandlerUPP(EventHandler);
-  InstallApplicationEventHandler (handlerUPP,
-                                7, eventTypes,
-                                NULL, NULL);
-
-  AEInstallEventHandler(kCoreEventClass, kAEOpenDocuments, NewAEEventHandlerUPP(AEodoc), 0, false);
-
-  EventTypeSpec events[] = {
-		{ kEventClassWindow, kEventWindowClosed }
-  };
-
-  EventTypeSpec commands[] = {
-        { kEventClassCommand, kEventCommandProcess }
-  };
-
-  InstallWindowEventHandler (mainWin->mWindow, 
-		   NewEventHandlerUPP (MainWindowEventHandler), 
-		   GetEventTypeCount(events), events, 
-		   mainWin->mWindow, NULL);
-
-  InstallWindowEventHandler(mainWin->mWindow, 
-            NewEventHandlerUPP (MainWindowCommandHandler), 
-            GetEventTypeCount(commands), commands, 
-            mainWin->mWindow, NULL);
-
-
-//  mainWin->StartRecordingVideo("jon.mov", nil);
-  
-  // Call the event loop
-  RunApplicationEventLoopWithCooperativeThreadSupport();
-#endif
-    
-    // autorun if a filepath was set
-    AutoRunFromPath();
-
-    return(0);
-}
-
-extern "C" int beeb_end()
-{
-  fprintf(stderr, "Shutting Down ...\n");
-  
-  if (tlog) fclose(tlog);
-
-  SoundReset();
-
-#if 0 //ACH - symbolic hotkey
-  if (PopSymbolicHotKeyMode != NULL)
-  {
-     PopSymbolicHotKeyMode(token);
-  }
-#endif
-
-  return(0);
-} /* main */
 
 void WriteLog(const char *fmt, ...)
 {
