+++ upgradenotes.rtf
@@ -1,0 +1,31 @@
+{\rtf1\ansi\ansicpg1252\cocoartf2513
+\cocoatextscaling0\cocoaplatform0{\fonttbl\f0\fswiss\fcharset0 Helvetica;}
+{\colortbl;\red255\green255\blue255;}
+{\*\expandedcolortbl;;}
+\paperw12240\paperh15840\margl1440\margr1440\vieww9000\viewh8400\viewkind0
+\pard\tx720\tx1440\tx2160\tx2880\tx3600\tx4320\tx5040\tx5760\tx6480\tx7200\tx7920\tx8640\pardirnatural\partightenfactor0
+
+\f0\fs24 \cf0 Making the Windows version work inside my framework.\
+\
+: skip anything that is a bit awkward\
+  music5000\
+  econet\
+  beebwindx\
+  beebwinio\
+  beebwinprefs\
+  beebwinspeech\
+  FolderSelectDialog\
+  SelectKeyDialog\
+  UserPortBreakoutBox\
+\
+\
+: get the MAIN header working \
+  beebwin.h\
+  main.h\
+  6502core.h\
+  beebmem.h\
+\
+: separate off the Windows and NonWindows (i.e. portable) code\
+  someone else can check it on windows!\
+\
+}
